+++ src/main/kotlin/com/coder/toolbox/util/CoderProtocolHandler.kt
@@ -43,6 +43,8 @@ open class CoderProtocolHandler(
     suspend fun handle(
         uri: URI,
         shouldWaitForAutoLogin: Boolean,
+        markAsBusy: () -> Unit,
+        unmarkAsBusy: () -> Unit,
         reInitialize: suspend (CoderRestClient, CoderCLIManager) -> Unit
     ) {
         val params = uri.toQueryParameters()
@@ -62,16 +64,27 @@ open class CoderProtocolHandler(
         val workspaceName = resolveWorkspaceName(params) ?: return
         val restClient = buildRestClient(deploymentURL, token) ?: return
         val workspace = restClient.workspaces().matchName(workspaceName, deploymentURL) ?: return
-        if (!prepareWorkspace(workspace, restClient, workspaceName, deploymentURL)) return
-
-        // we resolve the agent after the workspace is started otherwise we can get misleading
-        // errors like: no agent available while workspace is starting or stopping
-        val agent = resolveAgent(params, workspace) ?: return
-        if (!ensureAgentIsReady(workspace, agent)) return
 
         val cli = configureCli(deploymentURL, restClient)
         reInitialize(restClient, cli)
 
+        var agent: WorkspaceAgent
+        try {
+            markAsBusy()
+            context.refreshMainPage()
+            if (!prepareWorkspace(workspace, restClient, workspaceName, deploymentURL)) return
+            // we resolve the agent after the workspace is started otherwise we can get misleading
+            // errors like: no agent available while workspace is starting or stopping
+            // we also need to retrieve the workspace again to have the latest resources (ex: agent)
+            // attached to the workspace.
+            agent = resolveAgent(
+                params,
+                restClient.workspace(workspace.id)
+            ) ?: return
+            if (!ensureAgentIsReady(workspace, agent)) return
+        } finally {
+            unmarkAsBusy()
+        }
         val environmentId = "${workspace.name}.${agent.name}"
         context.showEnvironmentPage(environmentId)
 
@@ -173,7 +186,11 @@ open class CoderProtocolHandler(
                 }
 
                 try {
-                    restClient.startWorkspace(workspace)
+                    if (workspace.outdated) {
+                        restClient.updateWorkspace(workspace)
+                    } else {
+                        restClient.startWorkspace(workspace)
+                    }
                 } catch (e: Exception) {
                     context.logAndShowError(
                         CAN_T_HANDLE_URI_TITLE,
@@ -427,6 +444,7 @@ open class CoderProtocolHandler(
         )
     }
 }
+
 
 private fun CoderToolboxContext.popupPluginMainPage() {
     this.ui.showWindow()