+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - `Stop` action is now available for running workspaces that have an out of date template.
+- outdated and stopped workspaces are now updated and started when handling URI
 
 ## 0.3.0 - 2025-06-10